pyproject.toml
[tool.poetry]
name = "py-web-09"
version = "0.1.0"
description = ""
authors = ["Serghei <siracencoserghei@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
scrapy = "^2.11.1"
python-dotenv = "^1.0.1"
pymongo = "^4.6.2"
mongoengine = "^0.27.0"
tqdm = "^4.66.2"
faker = "^23.2.1"
redis = "^5.0.1"
redis-lru = "^0.1.2"
pika = "^1.3.2"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
